[refactor] BaseBoard 생성
