+++ resource-server/src/main/java/com/inhabas/api/domain/board/domain/NormalBoard.java
@@ -1,28 +1,22 @@
 package com.inhabas.api.domain.board.domain;
 
-import com.inhabas.api.domain.BaseEntity;
+import com.inhabas.api.domain.board.BaseBoard;
 import com.inhabas.api.domain.board.domain.valueObject.Contents;
 import com.inhabas.api.domain.board.domain.valueObject.Title;
 import com.inhabas.api.domain.comment.domain.Comment;
 import com.inhabas.api.domain.file.BoardFile;
 import com.inhabas.api.domain.member.domain.valueObject.MemberId;
-import com.inhabas.api.domain.menu.domain.valueObject.MenuId;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
-import javax.persistence.AttributeOverride;
 import javax.persistence.CascadeType;
-import javax.persistence.Column;
 import javax.persistence.DiscriminatorColumn;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.EntityListeners;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
 import javax.persistence.OneToMany;
@@ -40,31 +34,17 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @EntityListeners(AuditingEntityListener.class)
 @Inheritance(strategy = InheritanceType.JOINED)
 @DiscriminatorColumn(name = "TYPE", length = 15)
-public class NormalBoard extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Integer id;
-
-    @Embedded
-    protected Title title;
+public class NormalBoard extends BaseBoard {
 
     @Embedded
     protected Contents contents;
-
-    @Embedded
-    @AttributeOverride(name = "id", column = @Column(name = "menu_id"))
-    protected MenuId menuId;
-
-    @Embedded
-    @AttributeOverride(name = "id", column = @Column(name = "writer_id"))
-    protected MemberId writerId;
 
     @OneToMany(mappedBy = "parentBoard", cascade = CascadeType.ALL, orphanRemoval = true)
     protected List<Comment> comments = new ArrayList<>();
 
     @OneToMany(mappedBy = "parentBoard", cascade = CascadeType.ALL, orphanRemoval = true)
     protected Set<BoardFile> files = new HashSet<>();
+
 
     /* constructor */
 
@@ -82,53 +62,22 @@ public class NormalBoard extends BaseEntity {
 
     /* getter */
 
-    public Integer getId() {
-        return id;
-    }
-
-    public String getTitle() {
-        return title.getValue();
-    }
-
     public String getContents() {
         return contents.getValue();
-    }
-
-    public MenuId getMenuId() {
-        return menuId;
-    }
-
-    public MemberId getWriterId() {
-        return writerId;
     }
 
     public Set<BoardFile> getFiles() {
         return Collections.unmodifiableSet(files);
     }
 
+
     /* relation method */
-
-    @SuppressWarnings("unchecked")
-    public <T extends NormalBoard> T writtenBy(MemberId writerId){
-
-        if (Objects.isNull(this.writerId)) {
-            this.writerId = writerId;
-            return (T) this;
-        }
-        else {
-            throw new IllegalStateException("게시글 작성자를 수정할 수 없습니다.");
-        }
-    }
 
     public NormalBoard addFiles(Set<BoardFile> UploadFiles) {
         if (Objects.nonNull(UploadFiles))
             UploadFiles.forEach(this::addFile);
 
         return this;
-    }
-
-    public boolean isWriter(MemberId writerId) {
-        return this.writerId.equals(writerId);
     }
 
     public void addFile(BoardFile uploadFile) {
@@ -140,9 +89,4 @@ public class NormalBoard extends BaseEntity {
         comments.add(newComment);
     }
 
-    @SuppressWarnings("unchecked")
-    public <T extends NormalBoard> T inMenu(MenuId menuId) {
-        this.menuId = menuId;
-        return (T) this;
-    }
 }
